* Xtro.MDX.Utilities was advanced for Tutorial09.
Xtro.MDX:

* BufferSRV, ShaderResourceViewDescription, Texture1D_ArraySRV, Texture1D_SRV, Texture2DArraySRV, Texture2D_MS_ArraySRV, Texture2D_MS_SRV, Texture2D_SRV, Texture3D_SRV, TextureCubeSRV structs were added.
* SRV_Dimension and MeshFlag enums were added.
* All array<Byte> parameters were converted into UnmanagedMemory.
* CopyResource, CreateShaderResourceView, CreateShaderResourceView, CreateShaderResourceView, CreateShaderResourceView methods were added to Direct3D10.Device class.
* CreateMesh method was added to D3DX.
* ImageLoadInfo, TextureLoadInfo structs are now marshaled.
* Mesh and MeshBuffer classes were added.
* 64bit support implemented.
* UnmanagedMemory class is improved and now it has a generic version also.

+++ trunk/Libraries/Xtro.MDX/AdapterDescription.h
@@ -12,7 +12,7 @@ internal:
 		DedicatedVideoMemory = Native->DedicatedVideoMemory;
 		DedicatedSystemMemory = Native->DedicatedSystemMemory;
 		SharedSystemMemory = Native->SharedSystemMemory;
-		AdapterLuid = ((Int64)(Native->AdapterLuid.HighPart) << 32) | Native->AdapterLuid.LowPart;
+		AdapterLuid = ((long long)(Native->AdapterLuid.HighPart) << 32) | Native->AdapterLuid.LowPart;
 	}
 
 public:
@@ -24,7 +24,7 @@ public:
 	SIZE_T DedicatedVideoMemory;
 	SIZE_T DedicatedSystemMemory;
 	SIZE_T SharedSystemMemory;
-	Int64 AdapterLuid;
+	long long AdapterLuid;
 
 	static bool operator == (AdapterDescription Left, AdapterDescription Right)
 	{
